#183: improved logging

+++ cli/src/main/java/com/devonfw/tools/ide/tool/LocalToolCommandlet.java
@@ -66,6 +66,9 @@ public abstract class LocalToolCommandlet extends ToolCommandlet {
     VersionIdentifier installedVersion = getInstalledVersion();
     VersionIdentifier resolvedVersion = installation.resolvedVersion();
     if (resolvedVersion.equals(installedVersion)) {
+      IdeLogLevel level = silent ? IdeLogLevel.DEBUG : IdeLogLevel.INFO;
+      this.context.level(level).log("Version {} of tool {} is already installed", installedVersion,
+          getToolWithEdition());
       return false;
     }
     // we need to link the version or update the link.
@@ -132,12 +135,17 @@ public abstract class LocalToolCommandlet extends ToolCommandlet {
     FileAccess fileAccess = this.context.getFileAccess();
     if (Files.isDirectory(toolPath)) {
       if (Files.exists(toolVersionFile)) {
-        this.context.debug("Version {} of tool {} is already installed at {}", resolvedVersion,
-            getToolWithEdition(this.tool, edition), toolPath);
-        return createToolInstallation(toolPath, resolvedVersion, toolVersionFile);
+        if (this.context.isForceMode()) {
+          fileAccess.delete(toolPath);
+        } else {
+          this.context.debug("Version {} of tool {} is already installed at {}", resolvedVersion,
+              getToolWithEdition(this.tool, edition), toolPath);
+          return createToolInstallation(toolPath, resolvedVersion, toolVersionFile);
+        }
+      } else {
+        this.context.warning("Deleting corrupted installation at {}", toolPath);
+        fileAccess.delete(toolPath);
       }
-      this.context.warning("Deleting corrupted installation at {}", toolPath);
-      fileAccess.delete(toolPath);
     }
     Path target = toolRepository.download(this.tool, edition, resolvedVersion);
     fileAccess.mkdirs(toolPath.getParent());
